+++ src/helper/misc/Misc.js
@@ -17,10 +17,12 @@ class Misc extends Base {
    * @returns {Promise<Response<LinkMetadata>>}
    */
   async metadata (url) {
-    if (validator.isNullOrUndefined(url)) {
-      throw new WOLFAPIError('url cannot be null or empty', { url });
-    } else if (typeof url !== 'string') {
-      throw new WOLFAPIError('url must be type string', { url });
+    { // eslint-disable-line no-lone-blocks
+      if (validator.isNullOrUndefined(url)) {
+        throw new WOLFAPIError('url cannot be null or empty', { url });
+      } else if (typeof url !== 'string') {
+        throw new WOLFAPIError('url must be type string', { url });
+      }
     }
 
     if (this.metadataResults.some((result) => this.client.utility.string.isEqual(result.url, url))) {
@@ -61,8 +63,10 @@ class Misc extends Base {
    * @returns {Promise<Array<BlacklistLink>>}
    */
   async linkBlacklist (forceNew = false) {
-    if (!validator.isValidBoolean(forceNew)) {
-      throw new WOLFAPIError('forceNew must be a valid boolean', { forceNew });
+    { // eslint-disable-line no-lone-blocks
+      if (!validator.isValidBoolean(forceNew)) {
+        throw new WOLFAPIError('forceNew must be a valid boolean', { forceNew });
+      }
     }
 
     if (!forceNew && this.blacklist.length) {
@@ -78,11 +82,17 @@ class Misc extends Base {
 
   /**
    * Get the AWS security token
-   * @param {Boolean} requestNew
+   * @param {Boolean} forceNew
    * @returns {Promise<*>}
    */
-  async getSecurityToken (requestNew = false) {
-    if (!requestNew && this.client.cognito) {
+  async getSecurityToken (forceNew = false) {
+    { // eslint-disable-line no-lone-blocks
+      if (!validator.isValidBoolean(forceNew)) {
+        throw new WOLFAPIError('forceNew must be a valid boolean', { forceNew });
+      }
+    }
+
+    if (!forceNew && this.client.cognito) {
       return this.client.cognito;
     }
 
@@ -113,10 +123,12 @@ class Misc extends Base {
    * @returns {Promise<Response>}
    */
   async updateMessageSettings (messageFilterTier) {
-    if (!validator.isValidNumber(messageFilterTier)) {
-      throw new WOLFAPIError('messageFilterTier must be a valid number', { messageFilterTier });
-    } else if (!Object.values(MessageFilterTier).includes(parseInt(messageFilterTier))) {
-      throw new WOLFAPIError('messageFilterTier is not valid', { messageFilterTier });
+    { // eslint-disable-line no-lone-blocks
+      if (!validator.isValidNumber(messageFilterTier)) {
+        throw new WOLFAPIError('messageFilterTier must be a valid number', { messageFilterTier });
+      } else if (!Object.values(MessageFilterTier).includes(parseInt(messageFilterTier))) {
+        throw new WOLFAPIError('messageFilterTier is not valid', { messageFilterTier });
+      }
     }
 
     return await this.client.websocket.emit(