+++ src/models/DiscoverySection.js
@@ -11,13 +11,18 @@ class DiscoverySection extends Base {
     this.languageId = languageId;
     this.validity = data.validity ? new Validity(client, data?.validity) : undefined;
 
-    const elements = data.elementList;
+    const elements = [...new Set(data.elementList.map((element) => element.type))]
+      .reduce((result, value) => {
+        result[value] = data.elementList.filter((ele) => ele.type === value);
 
-    const sectionTitle = elements.find((element) => element.type === 'sectionTitle');
-    const heading = elements.find((element) => element.type === 'heading');
-    const images = elements.filter((element) => element.type === 'image');
-    const videos = elements.filter((element) => element.type === 'video');
-    const descriptions = elements.filter((element) => element.type === 'text');
+        return result;
+      }, {});
+
+    const sectionTitle = elements.sectionTitle[0];
+    const heading = elements.heading[0];
+    const images = elements.image;
+    const videos = elements.video;
+    const descriptions = elements.text;
 
     this.sectionTitle = sectionTitle?.properties?.text;
     this.title = heading?.properties?.text;
@@ -26,18 +31,18 @@ class DiscoverySection extends Base {
     this.videos = videos.length ? videos.map((video) => new TopicSectionVideo(client, video)) : undefined;
     this.additionalDescriptions = descriptions.length > 1 ? descriptions?.slice(1)?.map((description) => description?.properties?.text).filter(Boolean) : undefined;
 
-    const collection = elements.find((element) => element.type === 'collection');
+    const collection = elements.collection[0];
 
-    if (collection) {
-      const page = heading?.properties?.link?.url?.split('/').slice(-1)[0];
+    if (collection === undefined) { return false; }
 
-      (!fromSubPage && page)
-        ? this.page = page
-        : this.recipe = {
-          ...collection.properties.recipe,
-          type: collection.properties.type.replace('groupEvent', 'event')
-        };
-    }
+    const page = heading?.properties?.link?.url?.split('/').slice(-1)[0];
+
+    (!fromSubPage && page)
+      ? this.page = page
+      : this.recipe = {
+        ...collection.properties.recipe,
+        type: collection.properties.type.replace('groupEvent', 'event')
+      };
   }
 
   /**
@@ -50,17 +55,13 @@ class DiscoverySection extends Base {
       throw new WOLFAPIError(`${this.title} is not a collection section`, { page: this.toJSON() });
     }
 
-    if (this.page) {
-      const page = await this.client.discovery._getPage(this.page, this.languageId);
-
-      if (page.sections.length === 1 && page.sections[0].recipe) {
-        return await page.get(offset);
-      }
-
-      return page;
+    if (this.page === undefined) {
+      return await this.client.discovery._getAppropriateRecipeItems(this.recipe.id, this.languageId, this.recipe.max, offset, this.recipe.type);
     }
 
-    return await this.client.discovery._getAppropriateRecipeItems(this.recipe.id, this.languageId, this.recipe.max, offset, this.recipe.type);
+    const page = await this.client.discovery._getPage(this.page, this.languageId);
+
+    return page.sections.length === 1 && page.sections[0].recipe ? await page.get(offset) : page;
   }
 
   toJSON () {